Fix some bugs

+++ python_scripts/updater.py
@@ -144,8 +144,11 @@ class Updater:
             creator_db_hash = self.saver.get_creator_hash(creator)
             creator_ws_hash, comment = self.scanner.get_creator_hash(creator)
             if creator_ws_hash is None:
-                self.saver.delete_creator(creator)
-                creators_to_delete.append(creator)
+                if comment == "creator_not_found":
+                    self.saver.delete_creator(creator)
+                    creators_to_delete.append(creator)
+                else:
+                    continue
             elif creator_db_hash != creator_ws_hash:
                 creators_to_update.append(creator)
             else:
@@ -173,7 +176,12 @@ class Updater:
             script_db_hash = self.saver.get_script_hash(script[0], script[1])
             script_ws_hash, comment = self.scanner.get_script_hash(script[0], script[1])
             if script_ws_hash is None:
-                self.saver.delete_script(script[0], script[1])
+                if comment == "script_not_found":
+                    self.saver.delete_script(script[0], script[1])
+                else:
+                    continue
+            elif script_db_hash is None:
+                self.saver.insert_script(script[0], script[1])
             elif script_db_hash != script_ws_hash:
                 scripts_to_update.append(script)
             else:
